+++ CHANGELOG.md
@@ -15,10 +15,9 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `IntoIterator` implementations for `AlignmentStates` as well as an `iter()` method
 - Added `PartialEq` between `Cigar` and `AlignmentStates` without allocating on the heap
 - Added `score_from_path` for `GlobalPhmm`
-- Added `ref_len` and `query_len` to the `Alignment` struct
 - Added viterbi algorithm to `GlobalPhmm`
 - Added `LocalPhmm` and structs towards supporting other alignment modes in pHMMs
-
+- Added `ref_len` and `query_len` fields to the `Alignment` struct, as well as `to_reverse` and `make_reverse` methods
 
 ### Changed